Refactored `files` (IOS)
