+++ on_audio_query/ios/Classes/controllers/AudioController.swift
@@ -1,6 +1,6 @@
 import Flutter
 
-public class OnAudioController {
+public class AudioController {
     var call: FlutterMethodCall
     var result: FlutterResult
     
@@ -9,28 +9,27 @@ public class OnAudioController {
         self.result = result
     }
     
-    
     //This method will sort call according to request.
     public func chooseMethod() {
         // All necessary method to this plugin support both platforms, only playlists
         // are limited when using [IOS].
         switch call.method {
         case "querySongs":
-            OnAudioQuery(call: call, result: result).querySongs()
+            SongsQuery(call: call, result: result).querySongs()
         case "queryAlbums":
-            OnAlbumsQuery(call: call, result: result).queryAlbums()
+            AlbumsQuery(call: call, result: result).queryAlbums()
         case "queryArtists":
-            OnArtistsQuery(call: call, result: result).queryArtists()
+            ArtistsQuery(call: call, result: result).queryArtists()
         case "queryGenres":
-            OnGenresQuery(call: call, result: result).queryGenres()
+            GenresQuery(call: call, result: result).queryGenres()
         case "queryPlaylists":
-            OnPlaylistsQuery(call: call, result: result).queryPlaylists()
+            PlaylistsQuery(call: call, result: result).queryPlaylists()
         case "queryAudiosFrom":
-            OnAudiosFromQuery(call: call, result: result).queryAudiosFrom()
+            AudiosFromQuery(call: call, result: result).queryAudiosFrom()
         case "queryWithFilters":
-            OnWithFiltersQuery(call: call, result: result).queryWithFilters()
+            WithFiltersQuery(call: call, result: result).queryWithFilters()
         case "queryArtwork":
-            OnArtworkQuery(call: call, result: result).queryArtwork()
+            ArtworkQuery(call: call, result: result).queryArtwork()
         // The playlist for [IOS] is completely limited, the developer can only:
         //   * Create playlist
         //   * Add item to playlist (Unsuported, for now)
@@ -41,9 +40,9 @@ public class OnAudioController {
         //   * Remove item from playlist
         //   * Move item inside playlist
         case "createPlaylist":
-            OnPlaylistsController(call: call, result: result).createPlaylist()
+            PlaylistsController(call: call, result: result).createPlaylist()
         case "addToPlaylist":
-            OnPlaylistsController(call: call, result: result).addToPlaylist()
+            PlaylistsController(call: call, result: result).addToPlaylist()
         default:
             // All non suported methods will throw this error.
             result(FlutterMethodNotImplemented)
