+++ on_audio_query/ios/Classes/methods/helper/QueryHelper.swift
