+++ on_audio_query/ios/Classes/methods/queries/GenresQuery.swift
@@ -1,7 +1,7 @@
 import Flutter
 import MediaPlayer
 
-class OnAudioQuery {
+class GenresQuery {
     var args: [String: Any]
     var result: FlutterResult
     
@@ -11,16 +11,13 @@ class OnAudioQuery {
         self.result = result
     }
     
-    func querySongs() {
-        // The sortType.
-        let sortType = args["sortType"] as? Int ?? 0
-        
+    func queryGenres() {
         // Choose the type(To match android side, let's call "cursor").
-        let cursor = MPMediaQuery.songs()
-        // Using native sort from [IOS] you can only use the [Title], [Album] and
-        // [Artist]. The others will be sorted "manually" using [formatSongList] before
-        // send to Dart.
-        cursor.groupingType = checkSongSortType(sortType: sortType)
+        let cursor = MPMediaQuery.genres()
+        
+        // We don't need to define a sortType here. [IOS] only support
+        // the [Artist]. The others will be sorted "manually" using
+        // [formatSongList] before send to Dart.
         
         // This filter will avoid audios/songs outside phone library(cloud).
         let cloudFilter = MPMediaPropertyPredicate.init(
@@ -33,25 +30,29 @@ class OnAudioQuery {
         let hasPermission = SwiftOnAudioQueryPlugin().checkPermission()
         if hasPermission {
             // Query everything in background for a better performance.
-            loadSongs(cursor: cursor)
+            loadGenres(cursor: cursor.collections)
         } else {
             // There's no permission so, return empty to avoid crashes.
             result([])
         }
     }
     
-    private func loadSongs(cursor: MPMediaQuery!) {
+    private func loadGenres(cursor: [MPMediaItemCollection]!) {
         DispatchQueue.global(qos: .userInitiated).async {
-            var listOfSongs: [[String: Any?]] = Array()
+            var listOfGenres: [[String: Any?]] = Array()
             
-            // For each item(song) inside this "cursor", take one and "format"
+            // For each item(genre) inside this "cursor", take one and "format"
             // into a [Map<String, dynamic>], all keys are based on [Android]
             // platforms so, if you change some key, will have to change the [Android] too.
-            for song in cursor.items! {
-                // If the song file don't has a assetURL, is a Cloud item.
-                if !song.isCloudItem && song.assetURL != nil {
-                    let songData = loadSongItem(song: song)
-                    listOfSongs.append(songData)
+            for genre in cursor {
+                if !genre.items[0].isCloudItem && genre.items[0].assetURL != nil {
+                    var genreData = loadGenreItem(genre: genre)
+                    
+                    // Count and add the number of songs for every genre.
+                    let tmpMediaCount = getMediaCount(type: 0, id: genreData["_id"] as! UInt64)
+                    genreData["num_of_songs"] = tmpMediaCount
+                    
+                    listOfGenres.append(genreData)
                 }
             }
             
@@ -59,7 +60,7 @@ class OnAudioQuery {
             // inside the main thread).
             DispatchQueue.main.async {
                 // Here we'll check the "custom" sort and define a order to the list.
-                let finalList = formatSongList(args: self.args, allSongs: listOfSongs)
+                let finalList = formatGenreList(args: self.args, allGenres: listOfGenres)
                 self.result(finalList)
             }
         }
