+++ on_audio_query/ios/Classes/methods/queries/SongsQuery.swift
@@ -1,7 +1,7 @@
 import Flutter
 import MediaPlayer
 
-class OnPlaylistsQuery {
+class AudioQuery {
     var args: [String: Any]
     var result: FlutterResult
     
@@ -11,11 +11,16 @@ class OnPlaylistsQuery {
         self.result = result
     }
     
-    func queryPlaylists() {
-        // Choose the type(To match android side, let's call "cursor").
-        let cursor = MPMediaQuery.playlists()
+    func querySongs() {
+        // The sortType.
+        let sortType = args["sortType"] as? Int ?? 0
         
-        // TODO: Add sort type to [queryPlaylists].
+        // Choose the type(To match android side, let's call "cursor").
+        let cursor = MPMediaQuery.songs()
+        // Using native sort from [IOS] you can only use the [Title], [Album] and
+        // [Artist]. The others will be sorted "manually" using [formatSongList] before
+        // send to Dart.
+        cursor.groupingType = checkSongSortType(sortType: sortType)
         
         // This filter will avoid audios/songs outside phone library(cloud).
         let cloudFilter = MPMediaPropertyPredicate.init(
@@ -28,38 +33,34 @@ class OnPlaylistsQuery {
         let hasPermission = SwiftOnAudioQueryPlugin().checkPermission()
         if hasPermission {
             // Query everything in background for a better performance.
-            loadPlaylists(cursor: cursor.collections)
+            loadSongs(cursor: cursor)
         } else {
             // There's no permission so, return empty to avoid crashes.
             result([])
         }
     }
     
-    private func loadPlaylists(cursor: [MPMediaItemCollection]!) {
+    private func loadSongs(cursor: MPMediaQuery!) {
         DispatchQueue.global(qos: .userInitiated).async {
-            var listOfPlaylists: [[String: Any?]] = Array()
+            var listOfSongs: [[String: Any?]] = Array()
             
-            // For each item(playlist) inside this "cursor", take one and "format"
+            // For each item(song) inside this "cursor", take one and "format"
             // into a [Map<String, dynamic>], all keys are based on [Android]
             // platforms so, if you change some key, will have to change the [Android] too.
-            for playlist in cursor {
-                // If the first song file don't has a assetURL, is a Cloud item.
-                if !playlist.items[0].isCloudItem && playlist.items[0].assetURL != nil {
-                    var playlistData = loadPlaylistItem(playlist: playlist)
-                    
-                    // Count and add the number of songs for every genre.
-                    let tmpMediaCount = getMediaCount(type: 1, id: playlistData["_id"] as! UInt64)
-                    playlistData["num_of_songs"] = tmpMediaCount
-                    
-                    listOfPlaylists.append(playlistData)
+            for song in cursor.items! {
+                // If the song file don't has a assetURL, is a Cloud item.
+                if !song.isCloudItem && song.assetURL != nil {
+                    let songData = loadSongItem(song: song)
+                    listOfSongs.append(songData)
                 }
             }
             
             // After finish the "query", go back to the "main" thread(You can only call flutter
             // inside the main thread).
             DispatchQueue.main.async {
-                // TODO: Add sort type to [queryPlaylists].
-                self.result(listOfPlaylists)
+                // Here we'll check the "custom" sort and define a order to the list.
+                let finalList = formatSongList(args: self.args, allSongs: listOfSongs)
+                self.result(finalList)
             }
         }
     }
